api_fastapi/requirements.txt
fastapi
uvicorn
mysqlclient
bcrypt
python-jose
pydantic